refactor(modules): formalize discovery and my-scenes vertical slices
