+++ src/app/routes/AppRoutes.tsx
@@ -1,12 +1,12 @@
 import { Navigate, Route, Routes } from 'react-router-dom'
 import { Layout } from '@components/Layout'
+import { ScenesPage } from '@modules/discovery'
+import { MyScenesPage } from '@modules/my-scenes'
 import { SceneDetailPage } from '@modules/scene-detail'
 import { CreateScenePage } from '@pages/CreateScenePage'
 import { HomePage } from '@pages/HomePage'
 import { LoginPage } from '@pages/LoginPage'
-import { MyScenesPage } from '@pages/MyScenesPage'
 import { RegisterPage } from '@pages/RegisterPage'
-import { ScenesPage } from '@pages/ScenesPage'
 import { SettingsPage } from '@pages/SettingsPage'
 import { GuestOnlyRoute, ProtectedRoute } from './RouteGuards'
