+++ src/components/index.ts
@@ -9,6 +9,3 @@ export {
   ToggleField,
   Vector3Field,
 } from './SceneEditorControls'
-export { SceneCard } from './SceneCard'
-export { ScrollableTagBar } from './ScrollableTagBar'
-export { TagFilterBar } from './TagFilterBar'
