+++ src/modules/discovery/index.ts
@@ -1,0 +1,1 @@
+export { ScenesPage } from './ScenesPage'
