+++ src/modules/my-scenes/index.ts
@@ -1,0 +1,1 @@
+export { MyScenesPage } from './MyScenesPage'
