+++ src/pages/index.ts
@@ -1,9 +1,9 @@
 export { CreateScenePage } from './CreateScenePage'
 export { HomePage } from './HomePage'
 export { LoginPage } from './LoginPage'
-export { MyScenesPage } from './MyScenesPage'
+export { MyScenesPage } from '@modules/my-scenes'
 export { PlaceholderPage } from './PlaceholderPage'
 export { RegisterPage } from './RegisterPage'
 export { SceneDetailPage } from '@modules/scene-detail'
-export { ScenesPage } from './ScenesPage'
+export { ScenesPage } from '@modules/discovery'
 export { SettingsPage } from './SettingsPage'
